+++ lib/strict/attributes/configuration.rb
@@ -30,6 +30,7 @@ module Strict
       attr_reader :attributes
 
       def initialize(attributes:)
+        validate_instance_variables!(attributes)
         @attributes = attributes
         @attributes_index = attributes.to_h { |a| [a.name, a] }
       end
@@ -41,6 +42,20 @@ module Strict
       private
 
       attr_reader :attributes_index
+
+      def validate_instance_variables!(attributes)
+        attributes_by_instance_variable = {}
+        attributes.each do |attribute|
+          conflicting_attribute = attributes_by_instance_variable[attribute.instance_variable]
+          if conflicting_attribute
+            raise ArgumentError,
+                  "Attribute #{attribute.name.inspect} conflicts with #{conflicting_attribute.name.inspect} " \
+                  "because both use #{attribute.instance_variable}"
+          end
+
+          attributes_by_instance_variable[attribute.instance_variable] = attribute
+        end
+      end
     end
   end
 end