+++ spec/strict/attributes/configuration_spec.rb
@@ -33,4 +33,15 @@ RSpec.describe Strict::Attributes::Configuration do
       end.to raise_error(Strict::Attributes::Configuration::UnknownAttributeError)
     end
   end
+
+  it "rejects attributes that use the same backing instance variable" do
+    attributes = [
+      Strict::Attribute.make(:value?),
+      Strict::Attribute.make(:value!)
+    ]
+
+    expect do
+      described_class.new(attributes: attributes)
+    end.to raise_error(ArgumentError, /Attribute :value! conflicts with :value\?/)
+  end
 end